Refactor project structure to enforce module boundaries (Atlas)
- Move `src/grammar.rs` and `src/grammar.pest` to `src/parser/` to encapsulate parsing logic.
- Move `src/ast.rs` to `src/ast/mod.rs` for consistent directory-based modules.
- Move `src/codegen.rs` to `src/codegen/mod.rs` for consistent directory-based modules.
- Move `src/report.rs` to `src/tools/report.rs` as it is a CLI tool utility.
- Rename `src/parser.rs` to `src/parser/mod.rs`.
- Update `src/lib.rs` and documentation to reflect new paths (e.g., `glossa::parser::grammar`).
- Move regression test `tests/regression_operator_drop.rs` to `src/semantic/expressions.rs` to access internal types correctly.
- Update imports across the codebase.

This change improves cohesion by grouping related files and reducing top-level clutter in `src/`.
