+++ src/parser/grammar.rs
@@ -34,7 +34,7 @@
 //! # Example
 //!
 //! ```
-//! use glossa::grammar::parse;
+//! use glossa::parser::grammar::parse;
 //!
 //! let source = "«χαῖρε» λέγε.";
 //! let pairs = parse(source).unwrap();
@@ -50,7 +50,7 @@ use pest::Parser;
 use pest_derive::Parser;
 
 #[derive(Parser)]
-#[grammar = "grammar.pest"]
+#[grammar = "parser/grammar.pest"]
 pub struct GlossaParser;
 
 /// Parse a ΓΛΩΣΣΑ source string into a pest parse tree
