+++ src/tools/report.rs
